typecasting
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4392 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/VotingArbiter.cxx
@@ -129,7 +129,7 @@ bool VotingArbiter::closePoll(void)
   }
   // set starting time to exactly current time minus necessary vote time
   _startTime = TimeKeeper::getCurrent();
-  _startTime += -(_voteTime);
+  _startTime += -(float)(_voteTime);
 
   return true;
 }
